pyproject.toml
[project]
name = "kiso-docreader-mcp"
version = "0.1.0"
description = "Document text extraction (PDF/DOCX/XLSX/CSV/text) as a Model Context Protocol server"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.2",
    "pypdf>=4.0",
    "python-docx>=1.1",
    "openpyxl>=3.1",
]

[project.scripts]
kiso-docreader-mcp = "kiso_docreader_mcp.server:main"

[dependency-groups]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/kiso_docreader_mcp"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
